* extract single element container into a file
