+++ internal/containers/containers.go
@@ -78,31 +78,6 @@ func (s sliceContainer) InitWithSize(size int) {
 	}
 }
 
-type SingleElement struct {
-	dst             interface{}
-	assignableValue reflect.Value
-}
-
-func (s SingleElement) DstEl() interface{} {
-	return s.dst
-}
-
-func (s SingleElement) AddElementWithSubkey(_, _ string, value interface{}) {
-	s.AddElement("", value)
-}
-
-func (s SingleElement) AddElement(_ string, value interface{}) {
-	val := reflect.ValueOf(value)
-	assignableType := s.assignableValue.Type()
-	if assignableType.AssignableTo(val.Type()) {
-		s.assignableValue.Set(reflect.ValueOf(value))
-	} else {
-		s.assignableValue.Set(reflect.Indirect(val))
-	}
-}
-
-func (s SingleElement) InitWithSize(_ int) {}
-
 func NewContainer(dst interface{}) (Container, error) {
 	reflectValue := reflect.Indirect(reflect.ValueOf(dst))
 	var result Container
@@ -146,4 +121,3 @@ func NewContainer(dst interface{}) (Container, error) {
 
 var _ Container = mapContainer{}
 var _ Container = sliceContainer{}
-var _ Container = SingleElement{}
